# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API da Teste Receitas RG Sistemas",
        "description": "Documentação da API para o projeto da Teste Receitas RG Sistemas",
        "contact": {
            "email": "rafa.jefer@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/recipes": {
            "get": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Lista todas as receitas",
                "operationId": "aad95b03e882dfbecea7a0d7e0ec9c44",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de receitas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "1"
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Bolo de Cenoura"
                                            },
                                            "preparationTimeMinutes": {
                                                "type": "integer",
                                                "example": 60
                                            },
                                            "servings": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "ingredients": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            },
                                            "steps": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            },
                                            "userId": {
                                                "type": "string",
                                                "example": "1"
                                            },
                                            "userName": {
                                                "type": "string",
                                                "example": "Administrador"
                                            },
                                            "categoryId": {
                                                "type": "string",
                                                "example": null
                                            },
                                            "categoryName": {
                                                "type": "string",
                                                "example": null
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "example": [
                                    {
                                        "id": "1",
                                        "title": "Bolo de Cenoura",
                                        "preparationTimeMinutes": 60,
                                        "servings": 10,
                                        "ingredients": [
                                            "3 cenouras médias raladas",
                                            "4 ovos",
                                            "1 xícara de óleo",
                                            "2 xícaras de açúcar",
                                            "2 e 1/2 xícaras de farinha de trigo",
                                            "1 colher de sopa de fermento em pó"
                                        ],
                                        "steps": [
                                            "Bata no liquidificador as cenouras, ovos e óleo.",
                                            "Adicione o açúcar e bata novamente.",
                                            "Misture a farinha e o fermento em uma tigela e acrescente a mistura do liquidificador.",
                                            "Despeje em uma forma untada e asse em forno médio por 40 minutos."
                                        ],
                                        "userId": "1",
                                        "userName": "Administrador",
                                        "categoryId": "1",
                                        "categoryName": "Bolos e tortas doces"
                                    },
                                    {
                                        "id": "2",
                                        "title": "Pudim de Leite Condensado",
                                        "preparationTimeMinutes": 90,
                                        "servings": 8,
                                        "ingredients": [
                                            "1 lata de leite condensado",
                                            "2 latas de leite (medida da lata de leite condensado)",
                                            "3 ovos",
                                            "1 xícara de açúcar (para o caramelo)"
                                        ],
                                        "steps": [
                                            "Faça o caramelo com o açúcar e espalhe na forma.",
                                            "Bata os demais ingredientes no liquidificador.",
                                            "Despeje na forma caramelizada.",
                                            "Asse em banho-maria por cerca de 1h30."
                                        ],
                                        "userId": "1",
                                        "userName": "Administrador",
                                        "categoryId": null,
                                        "categoryName": null
                                    }
                                ]
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Cria uma nova receita",
                "operationId": "7882e720858cf0ca7fb47c7a2d36bdfe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "preparation_time_minutes": {
                                        "type": "integer"
                                    },
                                    "servings": {
                                        "type": "integer"
                                    },
                                    "ingredients": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "steps": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "title": "Novo Bolo de Cenoura",
                                "description": "Um bolo clássico, fofinho, delicioso e bem colorido.",
                                "ingredients": [
                                    "3 cenouras médias raladas",
                                    "4 ovos",
                                    "1 xícara de óleo",
                                    "2 xícaras de açúcar",
                                    "2 e 1/2 xícaras de farinha de trigo",
                                    "1 colher de sopa de fermento em pó"
                                ],
                                "steps": [
                                    "Bata no liquidificador as cenouras, ovos e óleo.",
                                    "Adicione o açúcar e bata novamente.",
                                    "Misture a farinha e o fermento em uma tigela e acrescente a mistura do liquidificador.",
                                    "Despeje em uma forma untada e asse em forno médio por 40 minutos."
                                ],
                                "preparation_time_minutes": 60,
                                "servings": 10,
                                "category_id": null
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Receita criada"
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}